#65 moving to pipelines.dwca package
